requirements.txt
pytest==9.0.1
python-dotenv==1.2.1
fastapi==0.118.3
uvicorn==0.38.0
requests==2.32.5
google-genai==1.52.0
google-adk==1.19.0
